Q2 and Q4 are updated.
Q2 works as described: sentenceCharCounts takes a sentence (e.g. ["i","love", "linux"]) and the helper function, sentenceCharCounts', also takes a sentence and returns the individual character counts of the each word. Then the  sentenceCharCounts function concatenates the sublists, which are character counts of individual words, then binds the tuples which have same key with fromListWith function.
Q4 works as described: changeKeys helper function changes the keys of the tuples, and dictWordsByCharCounts function binds the words according to their keys (a.k.a their character counts).