Q5 and Q7 are added.
In Q5, wordAnagrams function finds the anagram words of a given word. Output of Q4 is used to obtain character counts of lists of words. The character count of the given word is searched in the given list ([(CharacterCount, [Words])]) recursively. If the character count is equal to one of the character counts of list of words , then this list is returned as result.
In Q7, the character count of the 2nd list is subtracted from the character count of the 1st list.