Insert function is added.
In insert function, if the word is empty, then it means that we came to end of the word and the flag should be True.
Otherwise, the function would continue with the insert function of the Data.Map module. If the current character exists at trie at this level, then its children should be remained. Otherwise, the current character should be added to the trie without having any children.